Rename logging_config to logging
