+++ tests/test_logging.py
@@ -3,7 +3,7 @@ import sys
 from unittest.mock import MagicMock, patch
 
 import pytest
-from luthien_control.logging_config import (
+from luthien_control.logging import (
     DEFAULT_LOG_LEVEL,
     LOG_FORMAT,
     NOISY_LIBRARIES,
@@ -21,9 +21,9 @@ def reset_logging():
         root.handlers.clear()
 
 
-@patch("luthien_control.logging_config.Settings")
-@patch("luthien_control.logging_config.logging.basicConfig")
-@patch("luthien_control.logging_config.logging.getLogger")
+@patch("luthien_control.logging.Settings")
+@patch("luthien_control.logging.logging.basicConfig")
+@patch("luthien_control.logging.logging.getLogger")
 def test_setup_logging_default_level(mock_get_logger: MagicMock, mock_basic_config: MagicMock, MockSettings: MagicMock):
     """Test setup_logging uses DEFAULT_LOG_LEVEL when settings provide None or default."""
     # Arrange
@@ -46,14 +46,14 @@ def test_setup_logging_default_level(mock_get_logger: MagicMock, mock_basic_conf
         logger_mock.setLevel.assert_called_with(logging.WARNING)
 
     # Check the final confirmation log
-    mock_get_logger.assert_any_call("luthien_control.logging_config")
-    config_logger_mock = mock_get_logger("luthien_control.logging_config")
+    mock_get_logger.assert_any_call("luthien_control.logging")
+    config_logger_mock = mock_get_logger("luthien_control.logging")
     config_logger_mock.info.assert_called_once_with(f"Logging configured with level {DEFAULT_LOG_LEVEL}.")
 
 
-@patch("luthien_control.logging_config.Settings")
-@patch("luthien_control.logging_config.logging.basicConfig")
-@patch("luthien_control.logging_config.logging.getLogger")
+@patch("luthien_control.logging.Settings")
+@patch("luthien_control.logging.logging.basicConfig")
+@patch("luthien_control.logging.logging.getLogger")
 def test_setup_logging_specific_level(
     mock_get_logger: MagicMock, mock_basic_config: MagicMock, MockSettings: MagicMock
 ):
@@ -72,14 +72,14 @@ def test_setup_logging_specific_level(
     mock_settings_instance.get_log_level.assert_called_once_with(default=DEFAULT_LOG_LEVEL)
     mock_basic_config.assert_called_once_with(level=logging.DEBUG, format=LOG_FORMAT, stream=sys.stderr)
     # Check the final confirmation log
-    config_logger_mock = mock_get_logger("luthien_control.logging_config")
+    config_logger_mock = mock_get_logger("luthien_control.logging")
     config_logger_mock.info.assert_called_once_with(f"Logging configured with level {log_level_name}.")
 
 
-@patch("luthien_control.logging_config.Settings")
-@patch("luthien_control.logging_config.logging.basicConfig")
-@patch("luthien_control.logging_config.logging.getLogger")
-@patch("luthien_control.logging_config.print")  # Mock print to check warning
+@patch("luthien_control.logging.Settings")
+@patch("luthien_control.logging.logging.basicConfig")
+@patch("luthien_control.logging.logging.getLogger")
+@patch("luthien_control.logging.print")  # Mock print to check warning
 def test_setup_logging_invalid_level(
     mock_print: MagicMock, mock_get_logger: MagicMock, mock_basic_config: MagicMock, MockSettings: MagicMock
 ):
@@ -104,5 +104,5 @@ def test_setup_logging_invalid_level(
     # Check that basicConfig was called with the default level (INFO)
     mock_basic_config.assert_called_once_with(level=logging.INFO, format=LOG_FORMAT, stream=sys.stderr)
     # Check the final confirmation log used the default level name
-    config_logger_mock = mock_get_logger("luthien_control.logging_config")
+    config_logger_mock = mock_get_logger("luthien_control.logging")
     config_logger_mock.info.assert_called_once_with(f"Logging configured with level {DEFAULT_LOG_LEVEL}.")
